test: relocate the routine cases to agent/routine-scheduler.test.ts
Six cases move from `agent-service.routines.test.ts` to
`src/backend/agent/routine-scheduler.test.ts`: managing routines for yourself
and another agent through the six `openbot` tools, the cancellation marker
appended before an active run is deleted, the transition rollback that must
not leave a duplicate marker, the rejected invalid and cross-agent tool
mutations, the mutation rolled back when its transcript marker cannot
persist, and the queued work restored when a delete marker cannot persist.

Relocation only: each still constructs a real `AgentService` through
`agent-service-test-harness.ts` and asserts through public events, the
conversation markers and the JSONL protocol log. No assertion changed — the
diffs on both files touch only the import block and one stray blank line.
(Biome's unsafe fix renamed the now-unused `logPath` binding to `_logPath`,
which would have broken the destructuring; it is dropped instead.)

Fifteen cases stay in the shard because their consequence crosses a domain
the scheduler does not own. Named, since the shard is now the smaller half:
routine lifecycle markers must not add unread counts or search results
(conversation reads); a started routine delivery must keep running and issue
exactly one `turn/start` while its marker retries, and a terminal marker
retry must not stall turn completion (the queue engine and `#completeTurn`,
both of which the plan keeps in the core); routine approvals must stay
interactive while attention markers retry (the `RoutineAttention` seam, which
is two domains by construction). The rest of the shard — reactions,
response attachments, teammate relay, turn lifecycle — was never routine
work.

Twenty-one cases before, twenty-one after (15 + 6); all five agent-service
shards and all six `src/backend/agent/*.test.ts` files pass.
